Changes to interpreter_*.c to satisfy ESP32 C compiler
git-svn-id: https://svn.nfit.au.dk/svn/beta/trunk@18481 678b6752-a8a0-49ff-a511-7c2a3241fcd4

+++ MiniSystem/VM/interpreter_dump.c
@@ -380,7 +380,7 @@ void dumpCode(FILE *trace, ObjDesc desc){
 
 void dumpDesc(FILE *trace, int xdescNo) {
   ObjDesc desc;  
-  if ((desc = getDesc(xdescNo)) > 0 ) {
+  if ((int)(desc = getDesc(xdescNo)) > 0 ) {
     //fprintf(trace,"desc: %i ",desc);
     fprintf(trace,"\nClass ");
     //int i;
